installer: split probe.rs into probe and probe_parse
